requirements.txt
pillow
selenium
